requirements.txt
pillow
selenium
